Release/5.0.6.18

See merge request tetherfi_product/products/tmac/ui/agent-desktop!198

Former-commit-id: c31c52bda6a138d7f9292af81a27b9d990f20921

+++ src/app/modules/t-widgets/tw-content/twc-interaction/twc-interaction.component.ts
@@ -2,11 +2,13 @@ import { Component, ElementRef, OnDestroy, OnInit, ViewEncapsulation } from '@an
 import { AOTWidgetService } from '@services/aot-widget.service';
 import { TMACEventService } from '@services/tmac-event.service';
 import {
+    AutoCloseTabEvent,
     FaxReceivedEvent,
     GenericInteractionEvent,
     IncomingCallEvent,
     IncomingEmailEvent,
     InteractionClosedEvent,
+    IUIEvent,
     OutgoingCallEvent,
     OutgoingEmailEvent,
     TextChatIncomingEvent,
@@ -16,6 +18,7 @@ import { TWContentWrapper } from '@twidgets/utils/widget-wrapper/twc-wrapper';
 import { InteractionRef, InteractionWidgets, IWidget } from 'app/interfaces';
 import { ContentPageService } from 'app/services/content-page.service';
 import { InteractionManagerService } from 'app/services/interaction-manager.service';
+import { environment } from 'environments/environment';
 import { cloneDeep } from 'lodash';
 import { takeUntil } from 'rxjs/operators';
 
@@ -43,11 +46,6 @@ export class TwcInteractionComponent extends TWContentWrapper implements OnInit,
      * Type of content widget
      */
     type: string;
-
-    /**
-     * Is interaction flag
-     */
-    isInteraction: boolean;
 
     constructor(
         public hostElement: ElementRef,
@@ -108,9 +106,17 @@ export class TwcInteractionComponent extends TWContentWrapper implements OnInit,
         if (eventNames.length) {
             // subscribe to interaction events observable
             this._tmacEventService
-                .getConstructDisposeEvents([...eventNames, 'InteractionClosedEvent'])
+                .getConstructDisposeEvents<IUIEvent>([...eventNames, 'InteractionClosedEvent', 'AutoCloseTabEvent'])
                 .pipe(takeUntil(this.unsubscribeAll))
-                .subscribe((evts) => evts.forEach((evt) => this[evt.EventName](evt)));
+                .subscribe((evts) =>
+                    evts.forEach((evt) => {
+                        if (evt.EventName === 'InteractionClosedEvent' || evt.EventName === 'AutoCloseTabEvent') {
+                            this.tabCloseEvent(evt);
+                        } else {
+                            this[evt.EventName](evt);
+                        }
+                    })
+                );
 
             // subscribe to active interaction observable
             this._interactionManagerService.interactions.pipe(takeUntil(this.unsubscribeAll)).subscribe((interactions: InteractionRef[]) => {
@@ -137,28 +143,38 @@ export class TwcInteractionComponent extends TWContentWrapper implements OnInit,
      */
     private createWidgetList(evt: any, status: string, user: string, forceActive: boolean, otherData: any): void {
         // get the content widgets
-        const emailWidgets = cloneDeep(this.data.Data.Widgets) || [];
+        const widgets = cloneDeep(this.data.Data.Widgets) || [];
 
-        const staticWidgets = emailWidgets.Static || [];
-        const dynamicWidgets = (evt.WidgetConfigData && JSON.parse(evt.WidgetConfigData)) || emailWidgets.Dynamic || [];
-        const aotWidgets = emailWidgets.AOT || [];
+        const staticWidgets = widgets.Static ?? [];
+        const dynamicWidgets = ((environment.production && evt.WidgetConfigData && JSON.parse(evt.WidgetConfigData)) || widgets.Dynamic) ?? [];
+        const aotWidgets = widgets.AOT ?? [];
+
+        const routeOnInteraction = (forceActive || this.data.Data.RouteOnInteraction) ?? (['voice', 'textchat'].includes(this.type) ? true : false);
 
         // loop the widgets and add append interaction details
-        staticWidgets.forEach((widget: IWidget) => {
-            widget.InteractionDetails = evt;
-            widget.Data.Path = this.data.Data.Path;
-        });
+        staticWidgets
+            .filter((w: IWidget) => w.Config.Enabled)
+            .forEach((widget: IWidget) => {
+                widget.InteractionDetails = evt;
+                widget.Data.Path = this.data.Data.Path;
+                widget.Data.RouteOnInteraction = routeOnInteraction;
+            });
 
-        dynamicWidgets.forEach((widget: IWidget) => {
-            widget.InteractionDetails = evt;
-            widget.Data.Path = this.data.Data.Path;
-        });
+        dynamicWidgets
+            .filter((w: IWidget) => w.Config.Enabled)
+            .forEach((widget: IWidget) => {
+                widget.InteractionDetails = evt;
+                widget.Data.Path = this.data.Data.Path;
+                widget.Data.RouteOnInteraction = routeOnInteraction;
+            });
 
-        aotWidgets.forEach((widget: IWidget) => {
-            widget.ID = TUtils.Generic.uuid();
-            widget.InteractionDetails = evt;
-            widget.Data.Path = this.data.Data.Path;
-        });
+        aotWidgets
+            .filter((w: IWidget) => w.Config.Enabled)
+            .forEach((widget: IWidget) => {
+                widget.ID = TUtils.Generic.uuid();
+                widget.InteractionDetails = evt;
+                widget.Data.Path = this.data.Data.Path;
+            });
 
         // process aot widgets
         this._aotWidgetService.processAOTWidgets(aotWidgets);
@@ -178,7 +194,7 @@ export class TwcInteractionComponent extends TWContentWrapper implements OnInit,
             interactionId: evt.InteractionID,
             type: this.type,
             status: status,
-            isActive: this.interactions.length === 1 || forceActive,
+            isActive: this.interactions.length === 1 ?? forceActive,
             user: user || 'Customer',
             path: this.data.Data.Path,
             otherData: otherData
@@ -200,6 +216,10 @@ export class TwcInteractionComponent extends TWContentWrapper implements OnInit,
      * @param {OutgoingCallEvent} evt
      */
     OutgoingCallEvent(evt: OutgoingCallEvent): void {
+        // check if existing interaction and tab exist, then do not create the tab
+        if (evt.IsExistingInteraction && this.interactions.filter((i) => i.interactionId === evt.InteractionID)) {
+            return;
+        }
         this.createWidgetList(evt, 'outgoing', evt.PhoneNumber, true, {});
     }
 
@@ -243,26 +263,70 @@ export class TwcInteractionComponent extends TWContentWrapper implements OnInit,
     }
 
     /**
-     * To process interaction closed event for voice
+     * To process interaction closed event
      */
-    InteractionClosedEvent(evt: InteractionClosedEvent): void {
+    tabCloseEvent(evt: InteractionClosedEvent | AutoCloseTabEvent): void {
+        // since we have this widget for all interaction
+        // check this interaction id belongs to this widget interaction list
+        const thisInteraction = this.interactions.filter((i) => i.interactionId === evt.InteractionID);
+        if (!thisInteraction.length) {
+            return;
+        }
+
+        // remove the interaction reference
+        this._interactionManagerService.removeInteraction(evt.InteractionID);
+
         // close all the AOTs
-        this.interactions.forEach(i => {
+        this.interactions.forEach((i) => {
             if (i.interactionId === evt.InteractionID) {
-                i.widgets.aot.forEach(widget => {
+                i.widgets.aot.forEach((widget) => {
                     this._aotWidgetService.destroyWidget(widget.ID);
                 });
             }
         });
 
+        // get previous/next interaction index
+        const currentIndex = this.interactions.findIndex((i) => i.interactionId === evt.InteractionID);
+        const prevInteractionIndex = currentIndex - 1;
+        const nextInteractionIndex = currentIndex;
+
         // filter out the interaction
         this.interactions = this.interactions.filter((i: InteractionWidgets) => i.interactionId !== evt.InteractionID);
 
-        // if there are other item in the list auto select fist chat after closing current
+        // if there are other item in the list auto select fist interaction after closing current
         if (this.interactions.length > 0) {
-            this._interactionManagerService.updateInteraction(this.interactions[0].interactionId, {
+            // go to previous or next or first interaction
+            const routeInteraction = this.interactions[prevInteractionIndex] ?? this.interactions[nextInteractionIndex] ?? this.interactions[0];
+            this._interactionManagerService.updateInteraction(routeInteraction.interactionId, {
                 isActive: true
             });
+        } else {
+            this.activeInteraction = 0;
         }
     }
+
+    /**
+     * On page active callback
+     */
+    onActive = () => {
+        if (!this.activeInteraction) {
+            return;
+        }
+        // check if there is active interaction already
+        this._interactionManagerService.updateInteraction(this.activeInteraction, {
+            isActive: true
+        });
+    };
+
+    /**
+     * On page inactive callback
+     */
+    onInactive = () => {
+        if (!this.activeInteraction) {
+            return;
+        }
+        this._interactionManagerService.updateInteraction(this.activeInteraction, {
+            isActive: false
+        });
+    };
 }
